What about orientations?
WaitUntilAtTarget should also verify that the manipulandum is in the
correct orientation. Added parameters to the function to allow for
that. Modified the DexMouseApparatus to simulate rotations of the
manipulandum for testing purposes.

Modified WaitUntilAtTarget() to give a different message if the
manipulandum is not visible versus if it is not at the right position
or orientation.

Also updated some comments. Things that need to be modified are marked
with the tag 'TO DO'.