Extended Vector and Matrix operations.
Added functionality for inverting matrices, computing best-fit
transformations, etc. The program TestVectorsMixin.cpp extensively test
the vector and matrix code.

Note that I adopt the convention by which vectors are row vectors, not
column vectors. In this way, a matrix A[][] is in fact an array of row
vectors, making much of the code more straight forward.

Using row vectors, instead of column vectors, means that typically one
want to right multiply the vector by a matrix. Be careful not to forget
this.