Hardware Check Functionality
DexApparatus now performs checks on the hardware configuration based on
measurments of marker positions. The dialog boxes are still used in the
simulator, but they cause DexMouseTracker to report different marker
positions depending on the state of the check boxes and radio buttons.

DexApparatus no longer maintains a cached marker frame or a cached
manipulandum position or orientation. These values are computed as
needed. The full trajectory of the manipulandum and markers is,
however, computed and stored once at the end of a trial for use by the
post hoc tests.