Major update to ATI handling.
1. Updated the methods that compute forces and torques from the strain
gauges. The main modification is to rotate the sensor data so that both
sensors read forces in the same reference frame. That has implications
for the computation of net forces and torques.

2. Corrected the calculations of Center of Pressure (COP). Note that a
parameter is needed to define the minumum normal force required to
compute the center of pressure.

3. Cleaned up some other code to use instance variables, rather then
defines, for certain constants.