CheckForcePeaks()
This does not work well, but it shows the idea.
It may be better to check the acceleration rather than the force.
It could be a problem of the sampling rate that I am using here as
well. I will try samping at a fixed frequency, rather than polling and
interpolating.