Avoid degenerate case in computing manipulandum pose.
If the 4 or more markers fall in the same plane, we can get a
degenerate situation leading to a non-invertable matrix.
Here we shift the center of rotation to avoid the problem.
Note that this might not work in the most general case, but it should
work for all possible combinations of markers in the DEX case.
BUT IT NEEDS TO BE FURTHER TESTED as of today.