Clean up some issues with the compiler.
The compiler no longer creates the component objects. It just sets the
numbers of targets, sounds, etc. to the nominal values. If needed, we
will add a way to change those numbers in the command line.

Also added the data and time to the header of the compiled scripts.

Finally, I removed the automatic insertion of commands to set the
targets and tones off, since DEX does those automatically.