Directory reorganization.
Pictures are now in a directory just below DexterousManipulation, not
in the DexSimulatorApp directory.
Visual Studio should be reconfigured so that the simulator runs in the
DexOutput directory.
A makefile project has been added to make all the scripts at once. You
will need to get this project outside of GIT.