Simulate supine position.
Made extensive changes to the MouseTracker module so that we will more
accurately simulate the supine position. The markers on the bar and box
are transformed according to what was the configuration when the
alignment was done and what is the configuration now. The manipulandum
markers are made to move next to the target bar, no matter what the
configuration.