Explicit picture and message in SelectAndCheckConfiguration
My top level routine was automatically selecting the picture and
message to be put up, depending on the desired posture. But now the
desired posture must always be ‘upright’ even if sometimes we want
supine. So I give direct control over the message and picture to the
caller.

Some task routines that used this command have it commented out until
they can be updated.