CheckInstall()
Created a helper function that can be used to test if the configuration
is what we want. If it is already what is wanted, it should just
quietly move on. To be used in blocks after the first one of a protocol
in order to make sure that the configuration stays as it should.