Make mouse tracker to work for horizontal targets.
The mouse tracker simulator now places the manipulandum just to the
right of the box, not to the right of the target bar. This works for
the horizontal targets and since we never ask subjects to point to the
vertical targets when in the left position, it should work fine there,
too.