Sequence of events between tasks.
Added -sit option to indicate when the subject should be prompted to
sit or lie down and attach the belts.
Reinstated check of belts at the beginning of each task.
Adjusted some messages and changed the order of some preparation
actions.
Fixed a bug on retry following erroneous configuration detection
(affects simulator only).