Make sure the mass is good.
It’s possible to perform the experiment with no mass if one moves from
locked to unlocked position after the check is made. So we add an extra
check after the subject moves to the starting target. This will be
annoying if it actually happens, but most of the time the second test
should pass silently.