Options for proofs and a message list.
One can now enable or disable the generation of the proofs of the
message/picture combinations and one can redirect the output to a
specified directory (-proofs=).

DexLint now generates the list of messages and pictures even if the
proofs are not generated. This list serves as an input to
DexMessagePreviewer.